third_party/WebKit/LayoutTests/fast/events/mouse-event-source-device-event-sender-expected.txt

Test that mouse events from real mouse device should have a sourceDevice and its firesTouchEvents should be false.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


mousemove
PASS event.sourceDevice is non-null.
PASS event.sourceDevice.firesTouchEvents is false
mousedown
PASS event.sourceDevice is non-null.
PASS event.sourceDevice.firesTouchEvents is false
mouseup
PASS event.sourceDevice is non-null.
PASS event.sourceDevice.firesTouchEvents is false
click
PASS event.sourceDevice is non-null.
PASS event.sourceDevice.firesTouchEvents is false
mousedown
PASS event.sourceDevice is non-null.
PASS event.sourceDevice.firesTouchEvents is false
mouseup
PASS event.sourceDevice is non-null.
PASS event.sourceDevice.firesTouchEvents is false
click
PASS event.sourceDevice is non-null.
PASS event.sourceDevice.firesTouchEvents is false
dblclick
PASS event.sourceDevice is non-null.
PASS event.sourceDevice.firesTouchEvents is false
PASS successfullyParsed is true

TEST COMPLETE
